Moving GTFS vehicle type extractor to standalone lib file
